Refactor ServiceItem component file structure
